+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [Unreleased]
+
+### BREAKING CHANGES
+
+  -
+
+### Noteworthy
+
+  -
+
 # Changelog
 All notable changes to this project will be documented in this file.
 
@@ -12,7 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Noteworthy
 
-  -
+  - introduced `EntityRepository#paginate` method
 
 ## [4.3.0] - 2021-01-06